feature(tool-provider): renamed tool.py to provider.py and updated imports
