+++ treelang/ai/provider.py
@@ -76,6 +76,13 @@ class MCPToolProvider(ToolProvider):
 
 
 try:
+    """
+    Example of a non-MCP tool provider using LlamaIndex for tools provision.
+    In order to use this, you need to install LlamaIndex and have it available in your environment:
+    
+    `pip install llama-index`
+    
+    """
     from llama_index.tools import FunctionTool
 
     class LlamIndexToolProvider(ToolProvider):
